+++ src/oxl_ansible_webui/aw/config/defaults.py
@@ -59,6 +59,7 @@ CONFIG_DEFAULTS = {
     'path_ansible_config': _get_existing_ansible_config_file(),
     'path_ssh_known_hosts': _get_defaults_docker('path_ssh_known_hosts'),
     'debug': False,
+    'audit': False,
     'logo_url': 'img/logo.svg',
     'ssl_file_key': None,
     'ssl_file_crt': None,